packages/justhtml/justhtml-0.13.0.tar.gz/justhtml-0.13.0/tests/justhtml-tests/iframe_srcdoc.dat

#data
#iframe-srcdoc
<!DOCTYPE foo><p><table></table>
#errors
#document
| <!DOCTYPE foo>
| <html>
|   <head>
|   <body>
|     <p>
|     <table>
